market-data-stream-backend/requirements.txt
fastapi==0.115.5
Requests==2.32.3
uvicorn==0.32.1
websocket_client==1.8.0
